refactor(configure): extract setup_quality_gates into configurehandler (Option B)
First increment of the flat-package extraction the architecture review recommended:
drain domain logic out of `cmd/browser-agent/tools_*.go` into the already-established
handler sub-packages, leaving thin `package main` entry points.

Move the entire setup_quality_gates handler (scaffolding of .kaboom.json /
kaboom-code-standards.md + Claude Code hook installation) into
`configurehandler.HandleSetupQualityGates`, following the existing Deps pattern:
- new `Deps.GetActiveCodebase()` + adapter method on *ToolHandler
- package-main helpers translated to the sub-package equivalents
  (parseArgs→lenientUnmarshal, ErrX→mcp.ErrX, withParam→mcp.WithParam, fail/succeed)
- `tools_configure_quality_gates.go`: 355 LOC -> a 16-line delegator
- registry wiring unchanged (entry method keeps its signature)

Behavior-preserving. Verified: go build ./... + go vet clean; gofmt clean;
configurehandler 86 tests pass; configure integration suite 210 tests pass.
(Also dropped a redundant duplicate `strings.Contains(cmd, "kaboom-hooks")` term.)
